// src/ts/constant/api.ts
export enum ApiPath {
    GetActuatorArchaius = '/actuator/archaius',
    GetActuatorAuditevents = '/actuator/auditevents',
    GetActuatorBeans = '/actuator/beans',
    GetActuatorCaches = '/actuator/caches',
    DeleteActuatorCaches = '/actuator/caches',
    GetActuatorHealth = '/actuator/health',
    GetActuatorConditions = '/actuator/conditions',
    PostActuatorShutdown = '/actuator/shutdown',
    GetActuatorConfigprops = '/actuator/configprops',
    GetActuatorEnv = '/actuator/env',
    PostActuatorEnv = '/actuator/env',
    DeleteActuatorEnv = '/actuator/env',
    GetActuatorInfo = '/actuator/info',
    GetActuatorLogfile = '/actuator/logfile',
    GetActuatorLoggers = '/actuator/loggers',
    PostActuatorLoggers = '/actuator/loggers',
    GetActuatorHeapdump = '/actuator/heapdump',
    GetActuatorThreaddump = '/actuator/threaddump',
    GetActuatorPrometheus = '/actuator/prometheus',
    GetActuatorMetrics = '/actuator/metrics',
    GetActuatorScheduledtasks = '/actuator/scheduledtasks',
    GetActuatorHttptrace = '/actuator/httptrace',
    GetActuatorMappings = '/actuator/mappings',
    PostActuatorRefresh = '/actuator/refresh',
    GetActuatorFeatures = '/actuator/features',
    GetActuatorServiceRegistry = '/actuator/service-registry',
    PostActuatorServiceRegistry = '/actuator/service-registry',
    GetActuator = '/actuator',
    PostCacheClear = '/cache/clear',
    GetCache = '/cache',
    PostCacheEvict = '/cache/evict',
    GetDiscoveryInstance = '/discovery/instance',
    GetDiscoveryService = '/discovery/service',
    GetProxy = '/proxy',
    PostProxy = '/proxy',
    PutProxy = '/proxy',
    DeleteProxy = '/proxy',
    GetProxyContent = '/proxy/content',
    PostProxyContent = '/proxy/content',
    PutProxyContent = '/proxy/content',
    DeleteProxyContent = '/proxy/content',
    GetToolGenerateRsa = '/tool/generate/rsa',
    DeleteAccessLog = '/access/log',
    GetAccessLogPage = '/access/log/page',
    GetAccessLogGroupPage = '/access/log/group/page',
    GetAccessLogDetilPage = '/access/log/detil/page',
    PostApi = '/api',
    DeleteApi = '/api',
    GetApi = '/api',
    PutApi = '/api',
    PostEmailSend = '/email/send',
    DeleteExceptionLog = '/exception/log',
    GetExceptionLogPage = '/exception/log/page',
    PostPingerInit = '/pinger/init',
    PostPingerCheck = '/pinger/check',
    PutPublicRefresh = '/public/refresh',
    PostPublicFeature = '/public/feature',
    DeletePublicLogout = '/public/logout',
    GetPublicInit = '/public/init',
    PostPublicWow = '/public/wow',
    PostPublicLogin = '/public/login',
    PostRole = '/role',
    GetRolePage = '/role/page',
    GetRoleCombobox = '/role/combobox',
    PutRole = '/role',
    GetRolePermission = '/role/permission',
    PutRolePermission = '/role/permission',
    PostRoute = '/route',
    DeleteRoute = '/route',
    GetRoute = '/route',
    GetRouteFolder = '/route/folder',
    PutRoute = '/route',
    GetSelf = '/self',
    PutSelfPassword = '/self/password',
    PutSelfPasswordReset = '/self/password/reset',
    PutSelf = '/self',
    DeleteSession = '/session',
    GetSession = '/session',
    PostSubscription = '/subscription',
    DeleteSubscription = '/subscription',
    GetSubscriptionSelf = '/subscription/self',
    PutUserDisable = '/user/disable',
    PutUserEnable = '/user/enable',
    PostUser = '/user',
    PutUserLock = '/user/lock',
    PutUserUnlock = '/user/unlock',
    GetUserPage = '/user/page',
    PutUserPassword = '/user/password',
    GetUserCombobox = '/user/combobox',
    PutUserPasswordReset = '/user/password/reset',
    PutUser = '/user',
    DeleteScheduleLog = '/schedule/log',
    GetScheduleLog = '/schedule/log',
    GetSchedulePage = '/schedule/page',
    PostScheduleJob = '/schedule/job',
    DeleteScheduleJob = '/schedule/job',
    GetScheduleJobStatus = '/schedule/job/status',
    PutScheduleJobUnlock = '/schedule/job/unlock',
    GetScheduleJobKind = '/schedule/job/kind',
    PutScheduleJobStart = '/schedule/job/start',
    PostScheduleJobExecute = '/schedule/job/execute',
    PutScheduleJobStop = '/schedule/job/stop',
    PutScheduleJob = '/schedule/job',
}